created component for connector status/configuration
