+++ dynamic-mapping-ui/src/configuration/index.ts
@@ -1,3 +1,2 @@
 export * from './broker-configuration.module';
-export * from './shared/broker-configuration.service';
 export * from './shared/configuration.model';
