+++ dynamic-mapping-ui/src/connector/broker-connector.component.html
@@ -1,0 +1,76 @@
+<!--
+  ~ Copyright (c) 2022 Software AG, Darmstadt, Germany and/or Software AG USA Inc., Reston, VA, USA,
+  ~ and/or its subsidiaries and/or its affiliates and/or their licensors.
+  ~
+  ~ SPDX-License-Identifier: Apache-2.0
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ @authors Christof Strack
+  -->
+
+<c8y-title>Connector </c8y-title>
+<c8y-action-bar-item [placement]="'right'">
+  <!-- <button
+    class="btn btn-link"
+    title="{{ 'Reconnect notification endpoint' | translate }}"
+    (click)="clickedReconnect2NotificationEndpoint()"
+    [disabled]="!feature?.userHasMappingAdminRole"
+  >
+    <i c8yIcon="disconnected"></i>
+    {{ 'Reconnect notification endpoint' | translate }}
+  </button> -->
+
+  <button
+    class="btn btn-link"
+    title="{{ 'Refresh' | translate }}"
+    (click)="refresh()"
+    [disabled]="!feature?.userHasMappingAdminRole"
+  >
+    <i c8yIcon="refresh"></i>
+    {{ 'Refresh' | translate }}
+  </button>
+</c8y-action-bar-item>
+
+<div class="d-flex flex-wrap">
+  <div class="col-lg-6">
+    <div class="card">
+      <div class="card-header p-l-32 separator j-c-between">
+        <h4 class="card-title">
+          {{ 'Connector configuration' | translate }}
+        </h4>
+      </div>
+      <div class="card-block">
+        <div class="row">
+          <div class="col-lg-2">
+            <button
+              type="button"
+              name="addConfiguration"
+              id="addConfiguration"
+              class="btn btn-default btn-sm"
+              (click)="onConfigurationAdd()"
+              translate
+              [disabled]="!feature?.userHasMappingAdminRole"
+              title="Add configuration"
+            >
+              <i c8yIcon="plus-circle-o"></i>&nbsp;Add configuration
+            </button>
+          </div>
+        </div>
+		<d11r-mapping-connector-configuration></d11r-mapping-connector-configuration>
+      </div>
+
+      <d11r-mapping-connector-status></d11r-mapping-connector-status>
+    </div>
+  </div>
+</div>
