+++ dynamic-mapping-ui/src/connector/index.ts
@@ -1,0 +1,3 @@
+export * from './broker-connector.module';
+export * from '../shared/connector-configuration.service';
+
