+++ dynamic-mapping-ui/src/extension/grid/extension.component.html
@@ -26,7 +26,7 @@
   <c8y-breadcrumb-item
     [icon]="'c8y-modules'"
     [label]="'Extensions' | translate"
-    [path]="'sag-ps-pkg-dynamic-mapping/node2/extension'"
+    [path]="'sag-ps-pkg-dynamic-mapping/node3/extension'"
   >
     <!-- <c8y-breadcrumb-item [icon]="'c8y-modules'" [label]="'Extensions' | translate"> -->
   </c8y-breadcrumb-item>
