+++ dynamic-mapping-ui/src/extension/index.ts
@@ -1,0 +1,2 @@
+export * from './extension.service';
+
