+++ dynamic-mapping-ui/src/extension/properties/extension-properties.component.html
@@ -46,11 +46,6 @@
               *ngFor="
                 let entry of extension?.extensionEntries;
                 let i = index;
-                loadMore: 'show';
-                enableVirtualScroll: true;
-                virtualScrollElementSize: 40;
-                virtualScrollStrategy: 'fixed';
-                virtualScrollContainerHeight: 400
               "
               #li
             >
