+++ dynamic-mapping-ui/src/extension/properties/extension-properties.component.ts
@@ -18,21 +18,21 @@
  *
  * @authors Christof Strack
  */
-import { Component } from '@angular/core';
+import { AfterViewInit, Component } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
-import { IManagedObject } from '@c8y/client';
-import { gettext } from '@c8y/ngx-components';
-import { ExtensionService } from '../share/extension.service';
+import { ExtensionService } from '../extension.service';
 import { NODE3 } from '../../shared';
 
 @Component({
   selector: 'd11r-mapping-extension-properties',
   templateUrl: './extension-properties.component.html'
 })
-export class ExtensionPropertiesComponent {
+export class ExtensionPropertiesComponent implements AfterViewInit {
   extensionsEntryForm: FormGroup;
-  extension: IManagedObject;
+  extension = {
+    name: undefined
+  };
   isLoading: boolean = true;
   LINK = `sag-ps-pkg-dynamic-mapping/${NODE3}/extension`;
   breadcrumbConfig: { icon: string; label: string; path: string };
@@ -40,13 +40,10 @@ export class ExtensionPropertiesComponent {
   constructor(
     private activatedRoute: ActivatedRoute,
     private extensionService: ExtensionService
-  ) {
-    this.refresh();
-  }
+  ) {}
 
   async refresh() {
     await this.load();
-    this.setBreadcrumbConfig();
   }
 
   async load() {
@@ -58,14 +55,12 @@ export class ExtensionPropertiesComponent {
   async loadExtension() {
     const { id } = this.activatedRoute.snapshot.params;
     const result = await this.extensionService.getExtensionsEnriched(id);
-    this.extension = result[0];
+    this.extension = result[0] as any;
   }
 
-  private setBreadcrumbConfig() {
-    this.breadcrumbConfig = {
-      icon: 'c8y-modules',
-      label: gettext('Extensions'),
-      path: `sag-ps-pkg-dynamic-mapping/${NODE3}/extension`
-    };
+  ngAfterViewInit() {
+    setTimeout(async () => {
+      await this.load();
+    }, 0);
   }
 }
