+++ dynamic-mapping-ui/src/mapping-tree/tree.component.ts
@@ -20,9 +20,7 @@
  */
 import {
   AfterViewInit,
-  ChangeDetectorRef,
   Component,
-  NgZone,
   OnInit,
   ViewChild,
   ViewEncapsulation
