--- dynamic-mapping-ui/src/mapping/index.ts
@@ -1,4 +1,0 @@
-export * from './core/mapping.service';
-export * from './mapping.module';
-export * from './renderer/name.renderer.component';
-export * from './shared/mapping.model';
