+++ dynamic-mapping-ui/src/shared/connector-status/connector-status.component.html
@@ -1,0 +1,111 @@
+<!--
+  ~ Copyright (c) 2022 Software AG, Darmstadt, Germany and/or Software AG USA Inc., Reston, VA, USA,
+  ~ and/or its subsidiaries and/or its affiliates and/or their licensors.
+  ~
+  ~ SPDX-License-Identifier: Apache-2.0
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~      http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  ~ @authors Christof Strack
+  -->
+
+<div *ngIf="feature?.userHasMappingAdminRole">
+  <div class="legend form-block col-lg-12">Connector Logs</div>
+  <div class="card-block">
+	<c8y-form-group>
+	  <form class="form-inline">
+		<div class="c8y-select-wrapper button-bottom-margin">
+		  <select
+			class="form-control"
+			style="width: 180px"
+			[(ngModel)]="filterStatusLog.eventType"
+			name="eventType"
+		  >
+			<option [value]="'ALL'">ALL types</option>
+			<option
+			  [value]="StatusEventTypes.STATUS_CONNECTOR_EVENT_TYPE"
+			>
+			  Connector lifecycle
+			</option>
+			<option
+			  [value]="StatusEventTypes.STATUS_SUBSCRIPTION_EVENT_TYPE"
+			>
+			  Subscription events
+			</option>
+			<option
+			  [value]="StatusEventTypes.STATUS_NOTIFICATION_EVENT_TYPE"
+			>
+			  Notification lifecycle
+			</option>
+		  </select>
+		</div>
+		<div class="c8y-select-wrapper button-bottom-margin m-l-16">
+		  <select
+			class="form-control"
+			style="width: 180px"
+			[(ngModel)]="filterStatusLog.connectorIdent"
+			name="connectorIdent"
+		  >
+			<option [value]="'ALL'">ALL configurations</option>
+			<option [value]="t.ident" *ngFor="let t of configurations">
+			  {{ t.name }}
+			</option>
+		  </select>
+		</div>
+		<button
+		  type="submit"
+		  (click)="updateStatusLogs()"
+		  class="btn btn-default btn-sm m-b-8 m-l-16"
+		>
+		  Update
+		</button>
+	  </form>
+	</c8y-form-group>
+	<div style="height: 340px; overflow: auto">
+	  <!-- <c8y-ui-empty-state *ngIf="(statusLogs)?.length === 0"
+		[icon]="'event'"
+		[title]="'No logs found.' | translate"
+		[horizontal]="true"
+	  ></c8y-ui-empty-state> -->
+	  <div class="timeline-list list-condensed">
+		<div
+		  class="timeline-list-item d-flex"
+		  *ngFor="let event of statusLogs"
+		>
+		  <small class="timeline-item-date">
+			{{ event.date | date: 'dd.MM.yy hh:mm:ss' }}
+		  </small>
+		  <div class="timeline-item-content d-flex flex-wrap">
+			<div class="list-item-icon">
+			  <i c8yIcon="c8y-events"></i>
+			</div>
+			<div class="list-item-body">
+			  <div class="row">
+				<div class="col-lg-12">
+				  <label class="small" translate>Status&nbsp;</label>
+				  <span>{{ event.status }}</span>
+				  <label class="small" translate
+					>&nbsp;&nbsp;Connector&nbsp;</label
+				  >
+				  <span>{{ event.connectorName }}</span>
+				  <br />
+				  <span>{{ event.message }}</span>
+				</div>
+			  </div>
+			</div>
+		  </div>
+		</div>
+	  </div>
+	</div>
+  </div>
+</div>
