+++ dynamic-mapping-ui/src/shared/connector-status/connector-status.component.style.css
@@ -19,26 +19,6 @@
  * @authors Christof Strack
  */
 
-import { NgModule } from '@angular/core';
-import { CoreModule, hookRoute } from '@c8y/ngx-components';
-import { NODE3, SharedModule } from '../shared';
-import { BrokerConfigurationComponent } from './broker-configuration.component';
-import { BsDropdownModule } from 'ngx-bootstrap/dropdown';
-
-@NgModule({
-  declarations: [BrokerConfigurationComponent],
-  imports: [
-    CoreModule,
-    SharedModule,
-    BsDropdownModule.forRoot()
-  ],
-  exports: [],
-  providers: [
-    hookRoute({
-      path: `sag-ps-pkg-dynamic-mapping/${NODE3}/configuration`,
-      component: BrokerConfigurationComponent,
-      // canActivate: [AdminGuard]
-    })
-  ]
-})
-export class BrokerConfigurationModule {}
+.button-bottom-margin{
+  margin-bottom: 8px;
+}
